+++ src/parser.py
@@ -132,6 +132,7 @@ def create_parser(description='Zero2Neuro'):
     parser.add_argument('--data_file', type=str, default=None, help='Input data file')
     parser.add_argument('--data_files', nargs='+', type=str, default=None, help='Input data file list')
     #parser.add_argument('--data_table_merge', nargs='+', type=str, default=None, help='Table merge specification')   # removed
+    
     parser.add_argument('--data_inputs', nargs='+', type=str, default=None, help='Columns in the table that are inputs')
     parser.add_argument('--data_inputs_file_name', type=str, default=None, help='Column in the table that is a file name (image)')
     parser.add_argument('--data_inputs_file_type', type=str, default='image', help='File type for individual examples (image, plugin)')
@@ -140,8 +141,8 @@ def create_parser(description='Zero2Neuro'):
     parser.add_argument('--data_outputs_file_name', type=str, default=None, help='Column in the table that is a file name (image)')
     parser.add_argument('--data_outputs_file_type', type=str, default='image', help='File type for individual examples (image, plugin)')
 
-    parser.add_argument('--data_input_outputs_file_name', type=str, default=None, help='Column in the table that is a file name for both inputs and outputs')
-    parser.add_argument('--data_input_outputs_file_type', type=str, default='plugin', help='File type for individual input/output examples (plugin)')
+    parser.add_argument('--data_generic_file_name', type=str, default=None, help='Column in the table that is a file name for both inputs and outputs')
+    parser.add_argument('--data_generic_file_type', type=str, default='plugin', help='File type for individual input/output examples (plugin)')
 
     parser.add_argument('--data_tag_examples', nargs='+', type=str, default=None, help='Columns in the table that are to be used as example-wise tags.  These will be saved in the results pkl file.')
     parser.add_argument('--data_weights', type=str, default=None, help='Column in the table that are the sample weights')